restructure types
